renamed jcache/cache-remove-all to jcache/annotation-cache-remove-all, renamed service classes
