+++ jcache/annotation-cache-remove-all/pom.xml
@@ -8,5 +8,5 @@
         <relativePath>../pom.xml</relativePath>        
     </parent>
 
-    <artifactId>cache-remove-all</artifactId>
+    <artifactId>annotation-cache-remove-all</artifactId>
 </project>
