+++ jcache/annotation-cache-remove-all/src/test/resources/beans.xml
