+++ jcache/annotation-cache-remove-all/src/test/resources/jboss-deployment-structure.xml
